fzf: setup fzf if installed
Check if fzf is installed by checking if fzf binary is present
in its default installation directory ($HOME/.fzf/bin/fzf). We
can't use `which` here as fzf is not in the `$PATH`. That's one
of the purposes of `.fzf.bash`.

Also check for the presence of `.fzf.functions` and source it.
`.fzf.functions` will contain additional bash functions/aliases.