dev: refine git ctags hooks
Make `gentags` hooks (the default no-op ctags hook for new
projects) is more informative. If no language-specific ctags-hooks
has been set, and it is the default for new projects, a short
message indicating that ctags-hook is disabled, and where to set
it.
When a `ctags_cmd` has been set in `gentags`, no output is
produced if the hook exits with 0, otherwise a short message
indicating that something went wrong is printed, along with the
command to print manually to analyse why things went wrong. The
command is the `git ctags` alias.